# engine: return error if result contains duplicate label set