--- a/engine/engine.go
+++ b/engine/engine.go
@@ -372,6 +372,9 @@ func (q *compatibilityQuery) Exec(ctx context.Context) (ret *promql.Result) {
 	if err != nil {
 		return newErrResult(ret, err)
 	}
+	if containsDuplicateLabelSet(resultSeries) {
+		return newErrResult(ret, errors.New("vector cannot contain metrics with the same labelset"))
+	}
 
 	series := make([]promql.Series, len(resultSeries))
 	for i := 0; i < len(resultSeries); i++ {
@@ -482,6 +485,24 @@ func newErrResult(r *promql.Result, err error) *promql.Result {
 		r.Err = err
 	}
 	return r
+}
+
+func containsDuplicateLabelSet(series []labels.Labels) bool {
+	if len(series) <= 1 {
+		return false
+	}
+	var h uint64
+	buf := make([]byte, 0)
+	seen := make(map[uint64]struct{}, len(series))
+	for i := range series {
+		buf = buf[:0]
+		h, buf = series[i].HashWithoutLabels(buf)
+		if _, ok := seen[h]; ok {
+			return true
+		}
+		seen[h] = struct{}{}
+	}
+	return false
 }
 
 func (q *compatibilityQuery) Statement() parser.Statement { return nil }