@fpetkovski 2023-03-08
Maybe we can check if adding this has any performance implications. 

You can run `bench-new`  on main followed by `mv benchmarks/new.out benchmarks/main.out`. Then you can run `bench-new` on your branch and check the output of `benchstat benchmarks/main.out benchmarks/new.out`