@MichaHoffmann 2023-03-08
Ran the commands: 
```
$ benchstat benchmarks/main.out benchmarks/new.out
goos: linux
goarch: amd64
pkg: github.com/thanos-community/promql-engine/engine
cpu: 11th Gen Intel(R) Core(TM) i5-1135G7 @ 2.40GHz
                                                     │ benchmarks/main.out │         benchmarks/new.out         │
                                                     │       sec/op        │    sec/op     vs base              │
RangeQuery/vector_selector-8                                  76.42m ± ∞ ¹   76.45m ± ∞ ¹       ~ (p=1.000 n=5)
RangeQuery/sum-8                                              49.59m ± ∞ ¹   49.37m ± ∞ ¹       ~ (p=1.000 n=5)
RangeQuery/sum_by_pod-8                                       63.17m ± ∞ ¹   63.04m ± ∞ ¹       ~ (p=0.690 n=5)
RangeQuery/topk-8                                             53.04m ± ∞ ¹   53.54m ± ∞ ¹       ~ (p=1.000 n=5)
RangeQuery/bottomk-8                                          52.99m ± ∞ ¹   53.25m ± ∞ ¹       ~ (p=1.000 n=5)
RangeQuery/rate-8                                             117.9m ± ∞ ¹   119.0m ± ∞ ¹       ~ (p=0.310 n=5)
RangeQuery/sum_rate-8                                         93.46m ± ∞ ¹   91.17m ± ∞ ¹       ~ (p=0.310 n=5)
RangeQuery/sum_by_rate-8                                      114.2m ± ∞ ¹   116.2m ± ∞ ¹       ~ (p=0.548 n=5)
RangeQuery/quantile_with_variable_parameter-8                 103.9m ± ∞ ¹   102.8m ± ∞ ¹       ~ (p=0.151 n=5)
RangeQuery/binary_operation_with_one_to_one-8                 60.83m ± ∞ ¹   61.66m ± ∞ ¹       ~ (p=1.000 n=5)
RangeQuery/binary_operation_with_many_to_one-8                151.7m ± ∞ ¹   153.0m ± ∞ ¹       ~ (p=0.310 n=5)
RangeQuery/binary_operation_with_vector_and_scalar-8          113.5m ± ∞ ¹   115.3m ± ∞ ¹       ~ (p=0.222 n=5)
RangeQuery/unary_negation-8                                   82.00m ± ∞ ¹   82.26m ± ∞ ¹       ~ (p=0.841 n=5)
RangeQuery/vector_and_scalar_comparison-8                     113.3m ± ∞ ¹   110.6m ± ∞ ¹       ~ (p=0.151 n=5)
RangeQuery/positive_offset_vector-8                           75.82m ± ∞ ¹   74.98m ± ∞ ¹       ~ (p=0.841 n=5)
RangeQuery/at_modifier_-8                                     64.21m ± ∞ ¹   65.34m ± ∞ ¹       ~ (p=0.421 n=5)
RangeQuery/at_modifier_with_positive_offset_vector-8          63.50m ± ∞ ¹   64.60m ± ∞ ¹  +1.73% (p=0.008 n=5)
RangeQuery/clamp-8                                            137.0m ± ∞ ¹   135.3m ± ∞ ¹  -1.24% (p=0.016 n=5)
RangeQuery/clamp_min-8                                        129.0m ± ∞ ¹   127.5m ± ∞ ¹       ~ (p=0.310 n=5)
RangeQuery/complex_func_query-8                               166.5m ± ∞ ¹   167.4m ± ∞ ¹       ~ (p=0.151 n=5)
RangeQuery/func_within_func_query-8                           143.6m ± ∞ ¹   142.9m ± ∞ ¹       ~ (p=0.421 n=5)
RangeQuery/aggr_within_func_query-8                           143.4m ± ∞ ¹   143.3m ± ∞ ¹       ~ (p=0.690 n=5)
RangeQuery/histogram_quantile-8                               365.6m ± ∞ ¹   370.9m ± ∞ ¹       ~ (p=0.421 n=5)
RangeQuery/sort-8                                             80.62m ± ∞ ¹   80.59m ± ∞ ¹       ~ (p=0.841 n=5)
RangeQuery/sort_desc-8                                        79.98m ± ∞ ¹   80.93m ± ∞ ¹       ~ (p=0.690 n=5)
NativeHistograms/selector-8                                   664.7m ± ∞ ¹   675.0m ± ∞ ¹       ~ (p=1.000 n=5)
NativeHistograms/sum-8                                        839.1m ± ∞ ¹   839.6m ± ∞ ¹       ~ (p=1.000 n=5)
NativeHistograms/rate-8                                        4.125 ± ∞ ¹    4.044 ± ∞ ¹  -1.98% (p=0.032 n=5)
NativeHistograms/sum_rate-8                                    4.125 ± ∞ ¹    4.105 ± ∞ ¹       ~ (p=0.548 n=5)
NativeHistograms/histogram_sum-8                               1.028 ± ∞ ¹    1.031 ± ∞ ¹       ~ (p=1.000 n=5)
NativeHistograms/histogram_count-8                             1.022 ± ∞ ¹    1.022 ± ∞ ¹       ~ (p=1.000 n=5)
NativeHistograms/histogram_quantile-8                         848.1m ± ∞ ¹   846.7m ± ∞ ¹       ~ (p=1.000 n=5)
geomean                                                       171.6m         171.8m        +0.11%
¹ need >= 6 samples for confidence interval at level 0.95

                                                     │ benchmarks/main.out │         benchmarks/new.out          │
                                                     │        B/op         │     B/op       vs base              │
RangeQuery/vector_selector-8                                 36.94Mi ± ∞ ¹   36.94Mi ± ∞ ¹       ~ (p=0.690 n=5)
RangeQuery/sum-8                                             9.931Mi ± ∞ ¹   9.889Mi ± ∞ ¹       ~ (p=0.548 n=5)
RangeQuery/sum_by_pod-8                                      22.47Mi ± ∞ ¹   22.50Mi ± ∞ ¹       ~ (p=0.690 n=5)
RangeQuery/topk-8                                            11.07Mi ± ∞ ¹   11.04Mi ± ∞ ¹       ~ (p=1.000 n=5)
RangeQuery/bottomk-8                                         11.07Mi ± ∞ ¹   11.08Mi ± ∞ ¹       ~ (p=0.841 n=5)
RangeQuery/rate-8                                            39.73Mi ± ∞ ¹   39.80Mi ± ∞ ¹  +0.16% (p=0.032 n=5)
RangeQuery/sum_rate-8                                        12.82Mi ± ∞ ¹   12.74Mi ± ∞ ¹       ~ (p=0.222 n=5)
RangeQuery/sum_by_rate-8                                     24.90Mi ± ∞ ¹   24.66Mi ± ∞ ¹  -0.98% (p=0.032 n=5)
RangeQuery/quantile_with_variable_parameter-8                38.97Mi ± ∞ ¹   38.90Mi ± ∞ ¹       ~ (p=0.151 n=5)
RangeQuery/binary_operation_with_one_to_one-8                21.37Mi ± ∞ ¹   21.39Mi ± ∞ ¹  +0.06% (p=0.008 n=5)
RangeQuery/binary_operation_with_many_to_one-8               54.34Mi ± ∞ ¹   54.40Mi ± ∞ ¹  +0.10% (p=0.008 n=5)
RangeQuery/binary_operation_with_vector_and_scalar-8         39.97Mi ± ∞ ¹   39.94Mi ± ∞ ¹       ~ (p=1.000 n=5)
RangeQuery/unary_negation-8                                  38.16Mi ± ∞ ¹   38.23Mi ± ∞ ¹  +0.18% (p=0.008 n=5)
RangeQuery/vector_and_scalar_comparison-8                    39.45Mi ± ∞ ¹   39.49Mi ± ∞ ¹       ~ (p=0.690 n=5)
RangeQuery/positive_offset_vector-8                          35.53Mi ± ∞ ¹   35.57Mi ± ∞ ¹  +0.11% (p=0.008 n=5)
RangeQuery/at_modifier_-8                                    31.62Mi ± ∞ ¹   31.67Mi ± ∞ ¹  +0.15% (p=0.008 n=5)
RangeQuery/at_modifier_with_positive_offset_vector-8         31.45Mi ± ∞ ¹   31.49Mi ± ∞ ¹  +0.14% (p=0.008 n=5)
RangeQuery/clamp-8                                           37.30Mi ± ∞ ¹   37.34Mi ± ∞ ¹  +0.12% (p=0.016 n=5)
RangeQuery/clamp_min-8                                       37.25Mi ± ∞ ¹   37.30Mi ± ∞ ¹       ~ (p=0.095 n=5)
RangeQuery/complex_func_query-8                              40.04Mi ± ∞ ¹   40.07Mi ± ∞ ¹       ~ (p=0.421 n=5)
RangeQuery/func_within_func_query-8                          39.82Mi ± ∞ ¹   39.80Mi ± ∞ ¹       ~ (p=0.841 n=5)
RangeQuery/aggr_within_func_query-8                          39.81Mi ± ∞ ¹   39.84Mi ± ∞ ¹  +0.09% (p=0.008 n=5)
RangeQuery/histogram_quantile-8                              57.86Mi ± ∞ ¹   57.91Mi ± ∞ ¹       ~ (p=1.000 n=5)
RangeQuery/sort-8                                            36.92Mi ± ∞ ¹   36.96Mi ± ∞ ¹       ~ (p=0.151 n=5)
RangeQuery/sort_desc-8                                       36.93Mi ± ∞ ¹   36.97Mi ± ∞ ¹  +0.12% (p=0.008 n=5)
NativeHistograms/selector-8                                  739.5Mi ± ∞ ¹   739.4Mi ± ∞ ¹       ~ (p=0.690 n=5)
NativeHistograms/sum-8                                       889.8Mi ± ∞ ¹   889.9Mi ± ∞ ¹       ~ (p=0.421 n=5)
NativeHistograms/rate-8                                      2.680Gi ± ∞ ¹   2.680Gi ± ∞ ¹       ~ (p=0.841 n=5)
NativeHistograms/sum_rate-8                                  2.828Gi ± ∞ ¹   2.827Gi ± ∞ ¹       ~ (p=0.841 n=5)
NativeHistograms/histogram_sum-8                             740.9Mi ± ∞ ¹   740.7Mi ± ∞ ¹       ~ (p=0.095 n=5)
NativeHistograms/histogram_count-8                           740.8Mi ± ∞ ¹   740.8Mi ± ∞ ¹       ~ (p=0.222 n=5)
NativeHistograms/histogram_quantile-8                        890.2Mi ± ∞ ¹   890.0Mi ± ∞ ¹       ~ (p=0.548 n=5)
geomean                                                      66.55Mi         66.54Mi        -0.02%
¹ need >= 6 samples for confidence interval at level 0.95

                                                     │ benchmarks/main.out │         benchmarks/new.out         │
                                                     │      allocs/op      │  allocs/op    vs base              │
RangeQuery/vector_selector-8                                  114.0k ± ∞ ¹   113.6k ± ∞ ¹  -0.41% (p=0.008 n=5)
RangeQuery/sum-8                                              108.6k ± ∞ ¹   108.1k ± ∞ ¹  -0.43% (p=0.008 n=5)
RangeQuery/sum_by_pod-8                                       203.8k ± ∞ ¹   203.4k ± ∞ ¹  -0.22% (p=0.008 n=5)
RangeQuery/topk-8                                             115.5k ± ∞ ¹   114.3k ± ∞ ¹  -1.04% (p=0.008 n=5)
RangeQuery/bottomk-8                                          118.4k ± ∞ ¹   114.0k ± ∞ ¹  -3.72% (p=0.008 n=5)
RangeQuery/rate-8                                             138.1k ± ∞ ¹   137.6k ± ∞ ¹  -0.31% (p=0.008 n=5)
RangeQuery/sum_rate-8                                         132.7k ± ∞ ¹   132.2k ± ∞ ¹  -0.36% (p=0.008 n=5)
RangeQuery/sum_by_rate-8                                      227.8k ± ∞ ¹   227.3k ± ∞ ¹  -0.23% (p=0.008 n=5)
RangeQuery/quantile_with_variable_parameter-8                 583.7k ± ∞ ¹   582.8k ± ∞ ¹  -0.16% (p=0.008 n=5)
RangeQuery/binary_operation_with_one_to_one-8                 93.33k ± ∞ ¹   93.21k ± ∞ ¹  -0.13% (p=0.008 n=5)
RangeQuery/binary_operation_with_many_to_one-8                177.3k ± ∞ ¹   176.9k ± ∞ ¹  -0.25% (p=0.008 n=5)
RangeQuery/binary_operation_with_vector_and_scalar-8          118.3k ± ∞ ¹   117.8k ± ∞ ¹  -0.39% (p=0.008 n=5)
RangeQuery/unary_negation-8                                   126.0k ± ∞ ¹   125.6k ± ∞ ¹  -0.33% (p=0.008 n=5)
RangeQuery/vector_and_scalar_comparison-8                     115.2k ± ∞ ¹   114.8k ± ∞ ¹  -0.39% (p=0.008 n=5)
RangeQuery/positive_offset_vector-8                           88.07k ± ∞ ¹   87.78k ± ∞ ¹  -0.34% (p=0.008 n=5)
RangeQuery/at_modifier_-8                                     62.76k ± ∞ ¹   62.62k ± ∞ ¹  -0.22% (p=0.008 n=5)
RangeQuery/at_modifier_with_positive_offset_vector-8          56.76k ± ∞ ¹   56.62k ± ∞ ¹  -0.24% (p=0.008 n=5)
RangeQuery/clamp-8                                            117.8k ± ∞ ¹   117.4k ± ∞ ¹  -0.37% (p=0.008 n=5)
RangeQuery/clamp_min-8                                        117.4k ± ∞ ¹   117.0k ± ∞ ¹  -0.37% (p=0.008 n=5)
RangeQuery/complex_func_query-8                               122.0k ± ∞ ¹   121.6k ± ∞ ¹  -0.36% (p=0.008 n=5)
RangeQuery/func_within_func_query-8                           139.0k ± ∞ ¹   138.5k ± ∞ ¹  -0.35% (p=0.008 n=5)
RangeQuery/aggr_within_func_query-8                           139.0k ± ∞ ¹   138.5k ± ∞ ¹  -0.32% (p=0.008 n=5)
RangeQuery/histogram_quantile-8                               658.0k ± ∞ ¹   658.2k ± ∞ ¹  +0.03% (p=0.008 n=5)
RangeQuery/sort-8                                             114.0k ± ∞ ¹   113.6k ± ∞ ¹  -0.38% (p=0.008 n=5)
RangeQuery/sort_desc-8                                        114.0k ± ∞ ¹   113.6k ± ∞ ¹  -0.38% (p=0.008 n=5)
NativeHistograms/selector-8                                   9.643M ± ∞ ¹   9.643M ± ∞ ¹       ~ (p=0.690 n=5)
NativeHistograms/sum-8                                        11.08M ± ∞ ¹   11.08M ± ∞ ¹       ~ (p=0.841 n=5)
NativeHistograms/rate-8                                       34.15M ± ∞ ¹   34.15M ± ∞ ¹       ~ (p=0.238 n=5)
NativeHistograms/sum_rate-8                                   35.59M ± ∞ ¹   35.59M ± ∞ ¹       ~ (p=0.548 n=5)
NativeHistograms/histogram_sum-8                              9.646M ± ∞ ¹   9.646M ± ∞ ¹  -0.00% (p=0.008 n=5)
NativeHistograms/histogram_count-8                            9.646M ± ∞ ¹   9.646M ± ∞ ¹       ~ (p=0.841 n=5)
NativeHistograms/histogram_quantile-8                         11.09M ± ∞ ¹   11.09M ± ∞ ¹       ~ (p=0.310 n=5)
geomean                                                       377.1k         375.7k        -0.37%
¹ need >= 6 samples for confidence interval at level 0.95

```

the difference seems negligible, new.out seems to allocate fraction of a percent less even ( but the benchmark was not super scientific as it was run on my private machine )